Add files via upload
This is the main branch with mostly stock hardware. Only difference is that here it's set up to work with a dual extruder setup. This is because I will configure all my 3D printers to be dual color printers. I will update and add more features to this. Note that the stepper drivers used are the TMC2208 drivers. If you use any other drivers make sure to modify the configuration.h file.